+++ CHANGELOG.md
@@ -31,6 +31,8 @@
   component adapter as static subclasses of this single enhanced component class. This means that React4j generates
   a single java file for each component. The intent is to make it easier to understand what is happening underneath
   the covers.
+* **\[core\]** 💥 Replace `Component.setState(...)` with `Component.scheduleStateUpdate(...)` so that it
+  accurately reflects intent.
 * **\[core\]** 💥 Remove `Component.forceUpdate()` and replace it with the api `Component.scheduleRender(boolean force)`
   where `force=true` calls the underlying `NativeComponent.forceUpdate()` and `force=false` is equivalent
   to `setState({})`.